Adjust .js file for type definiton & Reduce `tsc` errors
